Update www service port
Introducing this function allows us to update the port the service is
running on after we create the service entry in SNMP table. This has
further benefit of retrieving the port number from ranch instead of
expecting the consumer to provide it for us.

This is mainly useful for automated tests where a port may be selected
randomly at start time.